Generating IR for global variables as well as primary & binary expressions. (#4)
* Added codegen functions to each ast node

* Fixed build process to link to clang libraries correctly

* Basic llvm module print option availible

* Returning llvm ir constants for int, float, char, and bools.

* Complete overhaul of primary expression parsing logic, as there was a major error involving the value. Also fixed the lexer to update boolean values. We can now create llvm ir for global vaiable declarations!

* Created new codegen namespace to reduce component coupling

* Codegen now supports declaration, definition, and reassignment of globals.

* General config housekeeping

* Made integers explicitly i64

* LLVM implementation of declaring, defining, and assigning global level variables all stored within a top level function

* Added valgrind option to dockerfile for quick identification of issues (SEGFAULTS!!!)

* Binary expressions now go to LLVM IR!

* Documentation galore

+++ include/lexer/lexer.h
@@ -65,7 +65,7 @@ namespace lexer {
      * 
      * @par A type definition using std::variant to store different values associated with tokens, that can store strings, integers, floats, characters, or booleans.
      */
-    typedef std::variant<std::string, int, float, char, bool> lexer_stored_values;
+    typedef std::variant<std::string, int64_t, float, char, bool> lexer_stored_values;
 
     /**
      * @par A vector that stores tokens lexed from the input stream.
@@ -85,7 +85,7 @@ namespace lexer {
     /**
      * @par Stores the current integer value from the input stream if applicable.
      */
-    extern int integer_value;
+    extern int64_t integer_value;
 
     /**
      * @par Stores the current float value from the input stream if applicable.
